Refactor APNS struct to be a single instance
Rather than having a different type for Request and Application, share the same type.
